requirements-mac.txt
# PyTorch for Mac (Apple Silicon)
torch>=2.0.0
torchvision
torchaudio

# Transformers and related libraries
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0

# PEFT for LoRA
peft>=0.7.0

# TRL for SFTTrainer
trl>=0.7.0

# Optional but recommended
sentencepiece>=0.1.99
protobuf>=3.20.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# CLI
typer[all]>=0.9.0
rich>=13.0.0

# Data collection
requests>=2.31.0
gitpython>=3.1.0
python-dotenv>=1.0.0

# HuggingFace Hub
huggingface-hub>=0.20.0
